requirements.txt
click
pygithub
pylama
